feat: Add language toggle widget and integrate into profile screen
- Implemented LangToggle widget for switching between Russian and Turkmen languages.
- Integrated LangToggle into ProfileScreen for improved user experience.
- Refactored ProfileScreen layout and adjusted styles for consistency.
- Updated AuthProvider to handle balancePoints as a double for better precision.

+++ lib/features/auth/auth_repository.dart
@@ -333,7 +333,10 @@ class AuthRepository {
     await prefs.setString('status', user['status'] ?? "pending");
     await prefs.setString('shop_address', user['address'] ?? "");
     await prefs.setDouble('rating', (user['rating'] ?? 0.0).toDouble());
-    await prefs.setInt('balance_points', user['balance_points'] ?? 0);
+    await prefs.setDouble(
+      'balance_points',
+      (user['balance_points'] ?? 0.0).toDouble(),
+    );
     await prefs.setDouble(
       'wallet_balance',
       (user['wallet_balance'] ?? 0.0).toDouble(),
